Rename files and take classes out of separate namespaces
