+++ CesiumGltf/include/CesiumGltf/PropertyType.h
@@ -6,6 +6,20 @@
 
 namespace CesiumGltf {
 enum class PropertyType {
+  Invalid,
+  Scalar,
+  Vec2,
+  Vec3,
+  Vec4,
+  Mat2,
+  Mat3,
+  Mat4,
+  String,
+  Boolean,
+  Enum
+};
+
+enum class PropertyComponentType {
   None,
   Int8,
   Uint8,
@@ -17,15 +31,25 @@ enum class PropertyType {
   Uint64,
   Float32,
   Float64,
-  Boolean,
-  Enum,
-  String,
-  Array,
 };
 
-std::string convertPropertyTypeToString(CesiumGltf::PropertyType type);
+std::string convertPropertyTypeToString(PropertyType type);
 
 PropertyType convertStringToPropertyType(const std::string& str);
 
-PropertyType convertOffsetStringToPropertyType(const std::string& str);
+std::string
+convertPropertyComponentTypeToString(PropertyComponentType componentType);
+
+PropertyComponentType
+convertStringToPropertyComponentType(const std::string& str);
+
+PropertyComponentType
+convertArrayOffsetTypeStringToPropertyComponentType(const std::string& str);
+
+PropertyComponentType
+convertStringOffsetTypeStringToPropertyComponentType(const std::string& str);
+
+bool isPropertyTypeVecN(PropertyType type);
+
+bool isPropertyTypeMatN(PropertyType type);
 } // namespace CesiumGltf
