+++ CesiumGltf/test/TestPropertyTexturePropertyView.cpp
@@ -1,4 +1,4 @@
-#include "CesiumGltf/StructuralMetadataPropertyTexturePropertyView.h"
+#include "CesiumGltf/PropertyTexturePropertyView.h"
 
 #include <catch2/catch.hpp>
 #include <gsl/span>
@@ -10,7 +10,6 @@
 #include <vector>
 
 using namespace CesiumGltf;
-using namespace CesiumGltf::StructuralMetadata;
 
 TEST_CASE(
     "Test PropertyTexturePropertyView on property with invalid texture index") {
